evaluation_api/requirements.txt
fastapi
uvicorn[standard]
aiohttp
pydantic
python-dotenv
